backend_new/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
google-genai
PyJWT==2.9.0